+++ src/Entity/Folder.php
@@ -1,0 +1,81 @@
+<?php
+
+namespace App\Entity;
+
+use ApiPlatform\Metadata\ApiResource;
+use App\Repository\FolderRepository;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Entity(repositoryClass: FolderRepository::class)]
+#[ORM\Table(name: "hevy_folders")]
+#[ApiResource]
+class Folder
+{
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
+
+    #[ORM\Column(length: 100)]
+    private ?string $title = null;
+
+    /**
+     * @var Collection<int, Routine>
+     */
+    #[ORM\OneToMany(targetEntity: Routine::class, mappedBy: 'folder')]
+    private Collection $routines;
+
+    public function __construct()
+    {
+        $this->routines = new ArrayCollection();
+    }
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getTitle(): ?string
+    {
+        return $this->title;
+    }
+
+    public function setTitle(string $title): static
+    {
+        $this->title = $title;
+
+        return $this;
+    }
+
+    /**
+     * @return Collection<int, Routine>
+     */
+    public function getRoutines(): Collection
+    {
+        return $this->routines;
+    }
+
+    public function addRoutine(Routine $routine): static
+    {
+        if (!$this->routines->contains($routine)) {
+            $this->routines->add($routine);
+            $routine->setFolder($this);
+        }
+
+        return $this;
+    }
+
+    public function removeRoutine(Routine $routine): static
+    {
+        if ($this->routines->removeElement($routine)) {
+            // set the owning side to null (unless already changed)
+            if ($routine->getFolder() === $this) {
+                $routine->setFolder(null);
+            }
+        }
+
+        return $this;
+    }
+}