+++ src/Entity/Routine.php
@@ -1,0 +1,114 @@
+<?php
+
+namespace App\Entity;
+
+use ApiPlatform\Metadata\ApiResource;
+use App\Repository\RoutineRepository;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Entity(repositoryClass: RoutineRepository::class)]
+#[ORM\Table(name: "hevy_routines")]
+#[ApiResource]
+class Routine
+{
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
+
+    #[ORM\Column(length: 255)]
+    private ?string $title = null;
+
+    #[ORM\ManyToOne(inversedBy: 'routines')]
+    private ?Folder $folder = null;
+
+    /**
+     * @var Collection<int, Exercise>
+     */
+    #[ORM\ManyToMany(targetEntity: Exercise::class, inversedBy: 'routines')]
+    private Collection $exercices;
+
+    #[ORM\Column]
+    private ?\DateTimeImmutable $createdAt;
+
+    #[ORM\Column]
+    private ?\DateTimeImmutable $updatedAt;
+
+    public function __construct()
+    {
+        $this->createdAt = $this->updatedAt = new \DateTimeImmutable("now", new \DateTimeZone("Europe/Paris"));
+        $this->exercices = new ArrayCollection();
+    }
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getTitle(): ?string
+    {
+        return $this->title;
+    }
+
+    public function setTitle(string $title): static
+    {
+        $this->title = $title;
+
+        return $this;
+    }
+    
+    public function getFolder(): ?Folder
+    {
+        return $this->folder;
+    }
+    
+    public function setFolder(?Folder $folder): static
+    {
+        $this->folder = $folder;
+        
+        return $this;
+    }
+
+    /**
+     * @return Collection<int, Exercise>
+     */
+    public function getExercices(): Collection
+    {
+        return $this->exercices;
+    }
+
+    public function addExercice(Exercise $exercice): static
+    {
+        if (!$this->exercices->contains($exercice)) {
+            $this->exercices->add($exercice);
+        }
+
+        return $this;
+    }
+
+    public function removeExercice(Exercise $exercice): static
+    {
+        $this->exercices->removeElement($exercice);
+
+        return $this;
+    }
+
+    public function getCreatedAt(): ?\DateTimeImmutable
+    {
+        return $this->createdAt;
+    }
+
+    public function getUpdatedAt(): ?\DateTimeImmutable
+    {
+        return $this->updatedAt;
+    }
+
+    public function setUpdatedAt(): static
+    {
+        $this->updatedAt = new \DateTimeImmutable("now", new \DateTimeZone("Europe/Paris"));
+
+        return $this;
+    }
+}