+++ tests/ApiResource/ExerciseTest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Tests\ApiResource;
+
+use ApiPlatform\Symfony\Bundle\Test\ApiTestCase;
+use App\Entity\Exercise;
+
+class ExerciseTest extends ApiTestCase
+{
+    public function testGetCollection(): void
+    {
+        $response = static::createClient()->request('GET', '/api/exercises');
+
+        $this->assertResponseStatusCodeSame(200);
+        $this->assertMatchesResourceCollectionJsonSchema(Exercise::class);
+        $this->assertJsonContains(['@id' => '/api/exercises']);
+    }
+}