+++ tests/ApiResource/FolderTest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Tests\ApiResource;
+
+use ApiPlatform\Symfony\Bundle\Test\ApiTestCase;
+use App\Entity\Folder;
+
+class FolderTest extends ApiTestCase
+{
+    public function testGetCollection(): void
+    {
+        $response = static::createClient()->request('GET', '/api/folders');
+
+        $this->assertResponseStatusCodeSame(200);
+        $this->assertMatchesResourceCollectionJsonSchema(Folder::class);
+        $this->assertJsonContains(['@id' => '/api/folders']);
+    }
+}