+++ tests/ApiResource/RoutineTest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Tests\ApiResource;
+
+use ApiPlatform\Symfony\Bundle\Test\ApiTestCase;
+use App\Entity\Routine;
+
+class RoutineTest extends ApiTestCase
+{
+    public function testGetCollection(): void
+    {
+        $response = static::createClient()->request('GET', '/api/routines');
+
+        $this->assertResponseStatusCodeSame(200);
+        $this->assertMatchesResourceCollectionJsonSchema(Routine::class);
+        $this->assertJsonContains(['@id' => '/api/routines']);
+    }
+}